Fix proto generation for missing query client in v0.12.0
The root cause is that `third_party` protos in `ibc-go` includes contains `auth.proto` (which does not contain query client). The generated `.rs` file from `cosmos-sdk` was being overwritten by the one generated by `ibc-go`. Fixes #219
