+++ cosmos-sdk-proto/src/prost/cosmos-sdk/cosmos.base.store.v1beta1.rs
