+++ cosmos-sdk-proto/src/prost/cosmos-sdk/cosmos.capability.v1beta1.rs
