+++ cosmos-sdk-proto/src/prost/cosmos-sdk/cosmos.crypto.ed25519.rs
