+++ cosmos-sdk-proto/src/prost/cosmos-sdk/cosmos.mint.v1beta1.rs
