+++ cosmos-sdk-proto/src/prost/ibc-go/ibc.applications.interchain_accounts.v1.rs
@@ -1,37 +1,3 @@
-/// An InterchainAccount is defined as a BaseAccount & the address of the account owner on the controller chain
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct InterchainAccount {
-    #[prost(message, optional, tag="1")]
-    pub base_account: ::core::option::Option<super::super::super::super::cosmos::auth::v1beta1::BaseAccount>,
-    #[prost(string, tag="2")]
-    pub account_owner: ::prost::alloc::string::String,
-}
-/// InterchainAccountPacketData is comprised of a raw transaction, type of transaction and optional memo field.
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct InterchainAccountPacketData {
-    #[prost(enumeration="Type", tag="1")]
-    pub r#type: i32,
-    #[prost(bytes="vec", tag="2")]
-    pub data: ::prost::alloc::vec::Vec<u8>,
-    #[prost(string, tag="3")]
-    pub memo: ::prost::alloc::string::String,
-}
-/// CosmosTx contains a list of sdk.Msg's. It should be used when sending transactions to an SDK host chain.
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct CosmosTx {
-    #[prost(message, repeated, tag="1")]
-    pub messages: ::prost::alloc::vec::Vec<::prost_types::Any>,
-}
-/// Type defines a classification of message issued from a controller chain to its associated interchain accounts
-/// host
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
-#[repr(i32)]
-pub enum Type {
-    /// Default zero value enumeration
-    Unspecified = 0,
-    /// Execute a transaction on an interchain accounts host chain
-    ExecuteTx = 1,
-}
 /// Metadata defines a set of protocol specific data encoded into the ICS27 channel version bytestring
 /// See ICS004: <https://github.com/cosmos/ibc/tree/master/spec/core/ics-004-channel-and-packet-semantics#Versioning>
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -107,4 +73,38 @@ pub struct RegisteredInterchainAccount {
     pub port_id: ::prost::alloc::string::String,
     #[prost(string, tag="3")]
     pub account_address: ::prost::alloc::string::String,
+}
+/// InterchainAccountPacketData is comprised of a raw transaction, type of transaction and optional memo field.
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct InterchainAccountPacketData {
+    #[prost(enumeration="Type", tag="1")]
+    pub r#type: i32,
+    #[prost(bytes="vec", tag="2")]
+    pub data: ::prost::alloc::vec::Vec<u8>,
+    #[prost(string, tag="3")]
+    pub memo: ::prost::alloc::string::String,
+}
+/// CosmosTx contains a list of sdk.Msg's. It should be used when sending transactions to an SDK host chain.
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct CosmosTx {
+    #[prost(message, repeated, tag="1")]
+    pub messages: ::prost::alloc::vec::Vec<::prost_types::Any>,
+}
+/// Type defines a classification of message issued from a controller chain to its associated interchain accounts
+/// host
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
+#[repr(i32)]
+pub enum Type {
+    /// Default zero value enumeration
+    Unspecified = 0,
+    /// Execute a transaction on an interchain accounts host chain
+    ExecuteTx = 1,
+}
+/// An InterchainAccount is defined as a BaseAccount & the address of the account owner on the controller chain
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct InterchainAccount {
+    #[prost(message, optional, tag="1")]
+    pub base_account: ::core::option::Option<super::super::super::super::cosmos::auth::v1beta1::BaseAccount>,
+    #[prost(string, tag="2")]
+    pub account_owner: ::prost::alloc::string::String,
 }
