+++ cosmos-sdk-proto/src/prost/ibc-go/ibc.lightclients.solomachine.v1.rs
