+++ cosmos-sdk-proto/src/prost/ibc-go/ibc.lightclients.solomachine.v2.rs
